Create code oil.
This function uses async/await to handle the asynchronous nature of the fetch method. It sends a GET request to the Oil Price API and retrieves the JSON data returned by the API. The data variable is then used to extract the price of oil from the JSON response, which is returned by the function.

Note that this function assumes that the Oil Price API provides data in the format of an array containing a single object with the price of oil in the price field. You may need to adjust this function to work with different APIs depending on the structure of their data. Additionally, you will need to obtain an API key for the Oil Price API and include it in the request URL as a query parameter.